+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 ### Upcoming
+
+*Major*
+
+* Initial version that connects Slack with the werewolf binary. ([#1](https://github.com/hjwylde/werewolf/issues/1))